requirements.txt
numpy>=1.24.0
opencv-python>=4.8.0
PyQt6>=6.4.0
pillow>=9.5.0 